+++ packages/gatsby-theme-vtex/src/templates/product.tsx
@@ -15,6 +15,7 @@ import {
   ProductPageQueryQueryVariables,
 } from './__generated__/ProductPageQuery.graphql'
 import SEO from '../components/ProductPage/SEO'
+import { isBot } from '../utils/env'
 
 const loadBelowTheFold = () => import('../components/ProductPage/BelowTheFold')
